// upahead_lambda/rest-api/packages/core/src/db/types/public/Message.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

import { type ConversationId } from './Conversation';
import { type ColumnType, type Selectable, type Insertable, type Updateable } from 'kysely';

/** Identifier type for public.Message */
export type MessageId = number & { __brand: 'MessageId' };

/** Represents the table public.Message */
export default interface MessageTable {
  id: ColumnType<MessageId, MessageId | undefined, MessageId>;

  conversation_id: ColumnType<ConversationId, ConversationId | undefined, ConversationId>;

  direction: ColumnType<string, string, string>;

  content_type: ColumnType<string, string, string>;

  content: ColumnType<string, string, string>;

  timestamp: ColumnType<Date, Date | string | undefined, Date | string>;
}

export type Message = Selectable<MessageTable>;

export type NewMessage = Insertable<MessageTable>;

export type MessageUpdate = Updateable<MessageTable>;